Refactor error handling classes for improved clarity and consistency; remove redundant definitions and enhance documentation.
