+++ metobs_toolkit/backend_collection/dev_collection.py
@@ -1,7 +1,8 @@
 """ Goal is to define a decorator that copies the docstring
 of one funtion to another. Useful for wrapping functions as
 class methods."""
-
+import logging
+import inspect
 from typing import Callable, TypeVar, Any
 from typing_extensions import ParamSpec
 import re
